pyproject.toml
[project]
name = "lunchlog"
version = "1"
description = "Django Rest Framework Backend application that allows users to manage their daily lunch receipts."
authors = [
    {name = "Khum Pun"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "boto3 (>=1.40.20,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    # Test tools installed into the development image so tests can run inside the container
    "pytest (>=7.4.2,<8.0.0)",
    "pytest-django (>=4.5.2,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
